moved and refactored benchmarks for testing
